Final version tech bot (#39)
* chain = Chain() where?

* u .gi

* current status

 1. Реализован лог аут. Добавлена кнопка retreive chat history, чтобы по команде пользователя вся история выводилась на экран. Убрать историю - clear conversation. Весь функционал  в сайдбаре включается (загрузка файла, селектор) только если пользователь залогинен. Подправила путь для get history в файле settings.py (client), поменяла history на get_history

* fix upload in client

* update chat histrory

* fix double upload

* model sellection (#36)

* add google packages to project

* fix exist upload folder

* fix chat history

* Dev (#37)

* model sellection

* simplyfiing and cleaning

---------


* fix form

* x form

* final version

Я сделала изменения в файле pages/3_chat.py.  Когда я стянула сегодня ветку, не работал uploader, clear conversation и постоянно висела история запросов. Я написала код,  сделала пул реквест и записала видео, чтобы показать, что программа полностью работает и готова к защите. ВАЖНО! Убедительная просьба уважать работу друг друга. Я написала код в файле клиента 3_chat.py,  не меняйте его, пожалуйста

* alisa code

* final

Изменения:
 - чат сохраняется, при желании страницу можно почистить.
 - файлы в базу повторно не сохранялис, а перезаписвыались в папку uploads, поэтому на сервер подкорректировала путь uploader.py, где просто проверяется существует ли указанный файл и если существует, то возвращатся сообщение ALREADY_EXISTS и никакого пересохранения не происходит.
В остальном - все работает, как надо.

---------

+++ server/src/routes/uploader.py
@@ -28,8 +28,19 @@ async def upload_file(
     file_paths = []
 
     target_folder = "uploads"
-    file_path = os.path.join(target_folder, file.filename)
 
+    # Проверяем существование папки
+    if not os.path.exists(target_folder):
+        # Если папка не существует, создаем её
+        os.makedirs(target_folder)
+
+###################
+## In case if file already exists skip all the steps about upload and signal the response        
+    file_path = os.path.join(target_folder, file.filename)
+    if os.path.exists(file_path):
+        return("ALREADY_EXISTS")
+###################
+    
     with open(file_path, "wb") as buffer:
         buffer.write(await file.read())
         file_paths.append(buffer.name)
